+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v3.7.0] - 2021-08-20
+
 * FloorMap - Improve configuration error alerting.
 * FloorMap - Fix bug in handling config provided via settings vice js.
 * FloorMap - Make zone editing configurable.
@@ -162,7 +165,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Intial open source release
 
-[Unreleased]: https://github.com/gchq/stroom-visualisations-dev/compare/v3.6.1-2...HEAD
+[Unreleased]: https://github.com/gchq/stroom-visualisations-dev/compare/v3.7.0...HEAD
+[v3.7.0]: https://github.com/gchq/stroom-visualisations-dev/compare/v3.6.1-2...v3.7.0
 [v3.6.1-2]: https://github.com/gchq/stroom-visualisations-dev/compare/v3.6.1-1...v3.6.1-2
 [v3.6.1-1]: https://github.com/gchq/stroom-visualisations-dev/compare/v3.6.1...v3.6.1-1
 [v3.6.1]: https://github.com/gchq/stroom-visualisations-dev/compare/v3.6.0...v3.6.1